+++ packages/cheqd-blockchain-api/src/api/index.js
@@ -1,4 +1,5 @@
 import { AbstractApiProvider } from '@docknetwork/credential-sdk/modules/abstract/common';
+import { GasPrice } from '@cosmjs/stargate';
 import {
   maybeToJSONString,
   fmtIterable,
@@ -15,6 +16,9 @@ import {
   FeemarketModule,
   // eslint-disable-next-line no-unused-vars
   CheqdSDK,
+  calculateDidFee,
+  CheqdQuerier,
+  setupFeemarketExtension,
 } from '@cheqd/sdk';
 import {
   DidRef,
@@ -53,7 +57,7 @@ import {
 } from '@docknetwork/credential-sdk/types';
 import { TypedEnum } from '@docknetwork/credential-sdk/types/generic';
 import pLimit from 'p-limit';
-import { buildTypeUrlObject, fullTypeUrl, fullTypeUrls } from './type-url';
+import { buildTypeUrlObject, fullTypeUrl } from './type-url';
 
 export class CheqdAPI extends AbstractApiProvider {
   #sdk;
@@ -83,13 +87,6 @@ export class CheqdAPI extends AbstractApiProvider {
   get sdk() {
     return this.ensureInitialized().#sdk;
   }
-
-  static Fees = buildTypeUrlObject(
-    DIDModule.fees.DefaultCreateDidDocFee,
-    DIDModule.fees.DefaultUpdateDidDocFee,
-    DIDModule.fees.DefaultDeactivateDidDocFee,
-    ResourceModule.fees.DefaultCreateResourceDefaultFee,
-  );
 
   static PayloadWrappers = buildTypeUrlObject(
     CheqdSetDidDocumentPayloadWithTypeUrlAndSignatures,
@@ -247,7 +244,27 @@ export class CheqdAPI extends AbstractApiProvider {
     };
     this.#sdk.signer.endpoint = options.endpoint; // HACK: cheqd SDK doesnt pass this with createCheqdSDK yet
 
+    if (options.gasPrice) {
+      this.setGasPrice(options.gasPrice);
+    } else {
+      const querier = (await CheqdQuerier.connectWithExtension(
+        rpcUrls[0],
+        setupFeemarketExtension,
+      ));
+
+      const feemarketModule = new FeemarketModule(this.#sdk.signer, querier);
+      const gasPrice = await feemarketModule.queryGasPrice('ncheq');
+      this.setGasPrice(gasPrice?.price ? GasPrice.fromString(`${gasPrice.price.amount}${gasPrice.price.denom}`) : undefined);
+    }
+
     return this;
+  }
+
+  /**
+   * Sets the gas price for fee estimation
+   */
+  setGasPrice(gasPrice) {
+    this.gasPrice = gasPrice ?? GasPrice.fromString('5000ncheq');
   }
 
   /**
@@ -287,18 +304,10 @@ export class CheqdAPI extends AbstractApiProvider {
    * @param {object|Array<object>} txOrTxs
    * @returns {BigInt}
    */
-  calculateFee(txOrTxs) {
-    const { Fees } = this.constructor;
-
-    const amount = fullTypeUrls(txOrTxs).reduce(
-      (total, typeUrl) => total + BigInt(Fees[typeUrl].amount),
-      0n,
-    );
-
-    return {
-      amount: String(amount),
-      denom: 'ncheq',
-    };
+  async calculateFee(txOrTxs, signerAddress, estimatedGas = null, multiplier = 1.3, memo = '') {
+    const gasEstimation = estimatedGas ?? (await this.simulate(signerAddress, Array.isArray(txOrTxs) ? txOrTxs : [txOrTxs], memo));
+    const usedFee = calculateDidFee(Math.round(gasEstimation * multiplier), this.gasPrice);
+    return usedFee.amount[0];
   }
 
   /**
@@ -395,9 +404,10 @@ export class CheqdAPI extends AbstractApiProvider {
     const sender = from ?? (await this.address());
     const txJSON = this.constructor.txToJSON(tx);
 
+    const estimatedGas = gas ?? (await this.estimateGas(tx, sender));
     const paymentObj = payment || {
-      amount: [].concat(fee ?? this.calculateFee(tx)),
-      gas: gas ?? (await this.estimateGas(tx, sender)),
+      amount: [].concat(fee ?? (await this.calculateFee(tx, sender, estimatedGas))),
+      gas: estimatedGas,
       payer: sender,
     };